[FEAT] User validation check for edit/delete posts
글 수정 페이지(GET), 글 삭제(DELETE)에 유저 유효성 검사 추가

+++ src/main/java/com/songpyeon/groupin/board/service/BoardService.java
@@ -5,6 +5,7 @@ import com.songpyeon.groupin.board.dto.BoardWriteDto;
 import com.songpyeon.groupin.board.repository.BoardRepository;
 import com.songpyeon.groupin.config.auth.PrincipalDetails;
 import com.songpyeon.groupin.handler.ex.CustomException;
+import com.songpyeon.groupin.handler.ex.CustomValidationException;
 import com.songpyeon.groupin.handler.ex.ErrorCode;
 import com.songpyeon.groupin.user.User;
 import lombok.RequiredArgsConstructor;
@@ -14,16 +15,22 @@ import org.springframework.security.core.parameters.P;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 import org.springframework.web.HttpMediaTypeNotAcceptableException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.validation.Valid;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 
 @RequiredArgsConstructor
@@ -86,15 +93,20 @@ public class BoardService {
     }
 
     //글 수정 페이지
-    public Board editPage(@PathVariable String category, @PathVariable int id, Board board) {
-        // user validation check 추가 필요
-
+    public Board editPage(@PathVariable String category, @PathVariable int id, PrincipalDetails principalDetails) {
         Board boardEntity;
         boardEntity = boardRepository.findByCategoryAndId(category, id);
+
         if (boardEntity == null) {
             throw new CustomException(ErrorCode.POSTS_NOT_FOUND);
         }
-        return boardEntity;
+
+        if (principalDetails.getUser().getUser_id() == boardEntity.getUser().getUser_id()) {
+            return boardEntity;
+        } else {
+            throw new CustomException(ErrorCode.NO_AUTHORITY);
+        }
+
     }
 
     // 글 수정하기
@@ -133,13 +145,17 @@ public class BoardService {
     }
 
     // 글 삭제하기
-    // user validation check 추가 필요
-    public void deletePost(@PathVariable String category, @PathVariable int id){
+    public void deletePost(@PathVariable String category, @PathVariable int id, PrincipalDetails principalDetails){
         Board post = boardRepository.findByCategoryAndId(category, id);
         if (post == null){
             throw new CustomException(ErrorCode.POSTS_NOT_FOUND);
         }
-        boardRepository.delete(post);
+
+        if (principalDetails.getUser().getUser_id() == post.getUser().getUser_id()) {
+            boardRepository.delete(post);
+        } else {
+            throw new CustomException(ErrorCode.NO_AUTHORITY);
+        }
     }
 
 }
